requirements.txt
streamlit
spacy
nltk
pdfplumber
python-docx
sentence-transformers
scikit-learn
openai
PyYAML
python-dotenv
matplotlib
reportlab
pytest
pandas
python-multipart
fuzzywuzzy
python-Levenshtein
tqdm
